Move fake row generator as a class.
