+++ bitsail-connectors/connector-kudu/pom.xml
@@ -66,6 +66,13 @@
             <version>${revision}</version>
             <scope>test</scope>
         </dependency>
+
+        <dependency>
+            <groupId>com.bytedance.bitsail</groupId>
+            <artifactId>connector-print</artifactId>
+            <version>${revision}</version>
+            <scope>test</scope>
+        </dependency>
     </dependencies>
 
     <build>
